feat: contact form
Refactored previously used input to newsletter specific input
