+++ src/components/Newsletter/index.ts
@@ -1,0 +1,3 @@
+import Newsletter from './Newsletter'
+
+export default Newsletter
