+++ src/components/NewsletterInput/index.ts
@@ -1,0 +1,3 @@
+import NewsletterInput from './NewsletterInput'
+
+export default NewsletterInput
